added smart filtering for private items
