+++ items/serializers/item_serializer.py
@@ -4,8 +4,7 @@ from inventory_logger.action_enum import ActionEnum
 from inventory_logger.utility.logger import LoggerUtility
 from items.factory.field_factory import FieldFactory
 from items.models import Item, Tag, Field
-from items.serializers.field_serializer import IntFieldSerializer, FloatFieldSerializer, \
-    ShortTextFieldSerializer, LongTextFieldSerializer
+from items.serializers.tag_serializer import NestedTagSerializer
 
 
 def create_fields(item):
@@ -18,12 +17,6 @@ class UniqueItemSerializer(serializers.ModelSerializer):
     class Meta:
         model = Item
         fields = ('id', 'name')
-
-
-class NestedTagSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Tag
-        fields = ('id', 'tag')
 
 
 class ItemSerializer(serializers.ModelSerializer):
@@ -49,25 +42,5 @@ class ItemSerializer(serializers.ModelSerializer):
         return item
 
 
-class DetailedItemSerializer(serializers.ModelSerializer):
-    tags = NestedTagSerializer(many=True, allow_null=True, required=False)
-    int_fields = IntFieldSerializer(many=True, read_only=True)
-    float_fields = FloatFieldSerializer(many=True, read_only=True)
-    short_text_fields = ShortTextFieldSerializer(many=True, read_only=True)
-    long_text_fields = LongTextFieldSerializer(many=True, read_only=True)
 
-    class Meta:
-        model = Item
-        fields = ('id', 'name', 'quantity', 'model_number', 'description', 'tags', 'int_fields' , 'float_fields',
-                  'short_text_fields', 'long_text_fields')
-
-    def update(self, instance, validated_data):
-        instance.name = validated_data.get('name', instance.name)
-        instance.quantity = validated_data.get('quantity', instance.quantity)
-        instance.model_number = validated_data.get('model_number', instance.model_number)
-        instance.description = validated_data.get('description', instance.description)
-        instance.save()
-        username = self.context['request'].user.username
-        LoggerUtility.log_as_system(ActionEnum.ITEM_MODIFIED, username + " Modified Item:" + instance.name)
-        return instance
