+++ items/serializers/tag_serializer.py
@@ -7,8 +7,15 @@ class TagSerializer(serializers.ModelSerializer):
         model = Tag
         fields = ('id', 'item', 'tag')
 
+
 class TagSingleSerializer(serializers.ModelSerializer):
     class Meta:
         model = Tag
         fields = ('tag', )
+
+
+class NestedTagSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Tag
+        fields = ('id', 'tag')
