+++ items/views/item_view.py
@@ -8,7 +8,8 @@ from inventory_logger.utility.logger import LoggerUtility
 from items.custom_pagination import LargeResultsSetPagination
 from items.logic.filter_item_logic import FilterItemLogic
 from items.models import Item
-from items.serializers.item_serializer import ItemSerializer, UniqueItemSerializer, DetailedItemSerializer
+from items.serializers.detailed_item_serializer import DetailedItemSerializer
+from items.serializers.item_serializer import ItemSerializer, UniqueItemSerializer
 
 
 class ItemList(generics.ListCreateAPIView):
